Merged stages
I changed all of our methods to return type pane. That way we have one main scene and stage rather than a bunch of stage windows.

Moving forward:
The filter food pane should be centered or turned into a toolbar.
OR
The nutritional information should get turned into a toolbar and moved to the bottom.
If anyone wants to peek at that, my addFood method is now a toolbar otherwise I'll look at it tomorrow!

LMK if any of this doesn't make sense or if you see a screw up with this!