Updated buttons and else field
Updated "Remove" button text field for meal list, also moved the "Analyze" button so it lines up with the side of the meal list pane. Also moved the foodTable.setItems(data); statement to outside of the else clause so it will update the food list table in both the if and else case.